If nats-config plug connected, get nats.cfg
The install hook now obtains a $SNAP_COMMON/server/nats.cfg file
from the nat-config content share, if it is connected.

The snap also now includes a getconfig app to get the config
after install, to make is easier for testing when the nats-config
interface is not previously connected (before snap install).